refactor(agent-service): 彻底删除旧目录 — orm/clients/config/middleware/utils/services 零残留
Import 迁移:
- app.middleware.chat_metrics → app.api.middleware (2 files)
- app.utils.middlewares.trace → app.api.middleware (10 files)
- app.services.content_parser → app.chat.content_parser (7 files)
- app.services.quick_search → app.chat.quick_search (1 file)
- app.clients/config → app.infra (sandbox_client, long_tasks)

删除 6 个旧目录 (3,579 行):
- app/orm/ — 旧 ORM 模型 + CRUD (已迁移到 data/)
- app/clients/ — 旧客户端 (已迁移到 infra/)
- app/config/ — 旧配置 (已迁移到 infra/config)
- app/middleware/ — 旧中间件 (已合并到 api/middleware)
- app/utils/ — 旧工具 (已分散到对应模块)
- app/services/ — 旧服务 (已迁移到 chat/)

测试: 272 passed, 0 failed
grep 旧 pattern: 零残留
最终代码量: 12,729 行
