--- apps/agent-service/app/clients/__init__.py
@@ -1,3 +1,0 @@
-"""
-Client modules for external services
-"""
