--- apps/agent-service/app/config/__init__.py
@@ -1,7 +1,0 @@
-"""
-配置模块，包含应用程序配置设置。
-"""
-
-from app.config.config import settings
-
-__all__ = ["settings"]
