--- apps/agent-service/app/config/logging_config.json
@@ -1,53 +1,0 @@
-{
-    "version": 1,
-    "disable_existing_loggers": false,
-    "formatters": {
-        "json": {
-            "()": "app.utils.logger.CustomJSONFormatter",
-            "fmt": "%(asctime)s %(levelname)s %(name)s %(message)s",
-            "datefmt": "%Y-%m-%d %H:%M:%S %z"
-        }
-    },
-    "handlers": {
-        "console": {
-            "class": "logging.StreamHandler",
-            "formatter": "json",
-            "stream": "ext://sys.stdout"
-        },
-        "file": {
-            "class": "logging.handlers.RotatingFileHandler",
-            "formatter": "json",
-            "filename": "/logs/agent-service/app.log",
-            "maxBytes": 10485760,
-            "backupCount": 5,
-            "mode": "a",
-            "encoding": "utf-8"
-        }
-    },
-    "loggers": {
-        "": {
-            "handlers": [
-                "console",
-                "file"
-            ],
-            "level": "INFO",
-            "propagate": true
-        },
-        "uvicorn": {
-            "handlers": [
-                "console",
-                "file"
-            ],
-            "level": "INFO",
-            "propagate": false
-        },
-        "uvicorn.error": {
-            "level": "INFO",
-            "handlers": [
-                "console",
-                "file"
-            ],
-            "propagate": false
-        }
-    }
-}
