--- apps/agent-service/app/orm/base.py
@@ -1,29 +1,0 @@
-from sqlalchemy import URL
-from sqlalchemy.ext.asyncio import AsyncSession, async_sessionmaker, create_async_engine
-
-from app.config.config import settings
-
-DATABASE_URL = URL.create(
-    "postgresql+asyncpg",
-    username=settings.postgres_user,
-    password=settings.postgres_password,
-    host=settings.postgres_host,
-    port=settings.postgres_port,
-    database=settings.postgres_db,
-    query={"ssl": "disable"},
-)
-
-engine = create_async_engine(
-    DATABASE_URL,
-    echo=False,
-    pool_pre_ping=True,
-    future=True,
-    pool_size=10,
-    pool_recycle=3600,
-)
-
-AsyncSessionLocal = async_sessionmaker(
-    bind=engine,
-    class_=AsyncSession,
-    expire_on_commit=False,
-)
