--- apps/agent-service/app/utils/__init__.py
@@ -1,18 +1,0 @@
-"""
-Utils module - includes utility functions and re-exports from inner_shared.
-"""
-
-# Local utilities
-# Re-export from inner_shared for convenience
-from inner_shared import (
-    BackfillWindow,
-    TimeRangeParser,
-    split_time,
-)
-
-__all__ = [
-    # From inner_shared
-    "BackfillWindow",
-    "TimeRangeParser",
-    "split_time",
-]
