--- apps/agent-service/app/utils/decorators/__init__.py
@@ -1,20 +1,0 @@
-"""
-Decorators module - re-exports from inner_shared with local extensions.
-"""
-
-# Re-export from inner_shared
-from inner_shared.decorators import (
-    dict_serialize,
-    handle_errors,
-    log_io,
-)
-
-# Keep local cache decorator that uses the app's Redis client
-from .cache_decorator import redis_cache
-
-__all__ = [
-    "handle_errors",
-    "dict_serialize",
-    "log_io",
-    "redis_cache",
-]
