--- apps/agent-service/app/utils/singleton/__init__.py
@@ -1,11 +1,0 @@
-"""
-Singleton module - re-exports from inner_shared.
-"""
-
-# Re-export from inner_shared
-from inner_shared.patterns import AsyncInitializable, AsyncSingletonFactory
-
-__all__ = [
-    "AsyncInitializable",
-    "AsyncSingletonFactory",
-]
